fixup! Move published_articles_by_tag to service
