+++ spec/services/articles/feeds/tag_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Articles::FindPublishedByTag, type: :service do
+RSpec.describe Articles::Feeds::Tag, type: :service do
   let(:tag) { "tag" }
   let!(:article) { create(:article) }
   let!(:tagged_article) { create(:article, tags: tag) }
